Backend/chatbot/requirements.txt
azure-functions
azure-core
azure-ai-formrecognizer
azure-identity
azure-cosmos
azure-search-documents
python-dotenv
requests
openai

